requirements.txt
requests>=2.25.1
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.3.0
beautifulsoup4>=4.9.3
lxml>=4.6.3